Macro lsio::repeat_color_with_ends

macro_rules! repeat_color_with_ends {
    ($color:expr, $e:expr, $text:expr, $lend:expr, $rend:expr, $size:expr) => { ... };
}

repeat_color_with_ends - Prints out a repeat of characters in a specific color and centers any text passed in along with printing left end and right end characters.

Currently prints characters using the color specified.

Example

repeat_color!(term::color::RED, "=", "", 80);Run